+++ examples/ai-core/src/stream-text/google-reasoning-fullstream.ts
@@ -1,38 +1,35 @@
-import { deepseek } from '@ai-sdk/deepseek';
+import { google } from '@ai-sdk/google';
 import { streamText } from 'ai';
 import 'dotenv/config';
 
 async function main() {
   const result = streamText({
-    model: deepseek('deepseek-reasoner'),
-    prompt: 'Invent a new holiday and describe its traditions.',
+    model: google('gemini-2.0-flash-thinking-exp'),
+    prompt: 'How many "r"s are in the word "strawberry"?',
   });
 
   let enteredReasoning = false;
   let enteredText = false;
   for await (const part of result.fullStream) {
+    if (part.type === 'error') {
+      console.error(part.error);
+      return;
+    }
+
     if (part.type === 'reasoning') {
       if (!enteredReasoning) {
         enteredReasoning = true;
-        console.log('\nSTREAMING REASONING:\n');
+        console.log('\nREASONING:\n');
       }
       process.stdout.write(part.textDelta);
     } else if (part.type === 'text-delta') {
       if (!enteredText) {
         enteredText = true;
-        console.log('\nSTREAMING TEXT:\n');
+        console.log('\nTEXT:\n');
       }
       process.stdout.write(part.textDelta);
     }
   }
-
-  console.log();
-  console.log('\nFINAL REASONING:\n', await result.reasoning);
-  console.log('\nFINAL TEXT:\n', await result.text);
-
-  console.log();
-  console.log('Token usage:', await result.usage);
-  console.log('Finish reason:', await result.finishReason);
 }
 
 main().catch(console.error);