+++ packages/ai/core/generate-text/tool-result.ts
@@ -1,23 +1,23 @@
-import { CoreTool } from '../tool';
 import { inferParameters } from '../tool/tool';
 import { ValueOf } from '../util/value-of';
+import { ToolSet } from './tool-set';
 
-export type { ToolResult } from '@ai-sdk/provider-utils';
+export type { CoreToolResult, ToolResult } from '@ai-sdk/provider-utils';
 
 // limits the tools to those with an execute value
-type ToToolsWithExecute<TOOLS extends Record<string, CoreTool>> = {
+type ToToolsWithExecute<TOOLS extends ToolSet> = {
   [K in keyof TOOLS as TOOLS[K] extends { execute: any } ? K : never]: TOOLS[K];
 };
 
 // limits the tools to those that have execute !== undefined
-type ToToolsWithDefinedExecute<TOOLS extends Record<string, CoreTool>> = {
+type ToToolsWithDefinedExecute<TOOLS extends ToolSet> = {
   [K in keyof TOOLS as TOOLS[K]['execute'] extends undefined
     ? never
     : K]: TOOLS[K];
 };
 
 // transforms the tools into a tool result union
-type ToToolResultObject<TOOLS extends Record<string, CoreTool>> = ValueOf<{
+type ToToolResultObject<TOOLS extends ToolSet> = ValueOf<{
   [NAME in keyof TOOLS]: {
     type: 'tool-result';
     toolCallId: string;
@@ -27,9 +27,16 @@ type ToToolResultObject<TOOLS extends Record<string, CoreTool>> = ValueOf<{
   };
 }>;
 
-export type ToolResultUnion<TOOLS extends Record<string, CoreTool>> =
-  ToToolResultObject<ToToolsWithDefinedExecute<ToToolsWithExecute<TOOLS>>>;
+export type ToolResultUnion<TOOLS extends ToolSet> = ToToolResultObject<
+  ToToolsWithDefinedExecute<ToToolsWithExecute<TOOLS>>
+>;
 
-export type ToolResultArray<TOOLS extends Record<string, CoreTool>> = Array<
+/**
+ * @deprecated Use `ToolResultUnion` instead.
+ */
+// TODO remove in v5
+export type CoreToolResultUnion<TOOLS extends ToolSet> = ToolResultUnion<TOOLS>;
+
+export type ToolResultArray<TOOLS extends ToolSet> = Array<
   ToolResultUnion<TOOLS>
 >;