+++ packages/ai/core/tool/tool.ts
@@ -36,7 +36,7 @@ This enables the language model to generate the input.
 
 The tool can also contain an optional execute function for the actual execution function of the tool.
  */
-export type CoreTool<PARAMETERS extends Parameters = any, RESULT = any> = {
+export type Tool<PARAMETERS extends Parameters = any, RESULT = any> = {
   /**
 The schema of the input that the tool expects. The language model will use this to generate the input.
 It is also used to validate the output of the language model.
@@ -91,31 +91,40 @@ The arguments for configuring the tool. Must match the expected arguments define
 );
 
 /**
+ * @deprecated Use `Tool` instead.
+ */
+// TODO remove in v5
+export type CoreTool<PARAMETERS extends Parameters = any, RESULT = any> = Tool<
+  PARAMETERS,
+  RESULT
+>;
+
+/**
 Helper function for inferring the execute args of a tool.
  */
 // Note: special type inference is needed for the execute function args to make sure they are inferred correctly.
 export function tool<PARAMETERS extends Parameters, RESULT>(
-  tool: CoreTool<PARAMETERS, RESULT> & {
+  tool: Tool<PARAMETERS, RESULT> & {
     execute: (
       args: inferParameters<PARAMETERS>,
       options: ToolExecutionOptions,
     ) => PromiseLike<RESULT>;
   },
-): CoreTool<PARAMETERS, RESULT> & {
+): Tool<PARAMETERS, RESULT> & {
   execute: (
     args: inferParameters<PARAMETERS>,
     options: ToolExecutionOptions,
   ) => PromiseLike<RESULT>;
 };
 export function tool<PARAMETERS extends Parameters, RESULT>(
-  tool: CoreTool<PARAMETERS, RESULT> & {
+  tool: Tool<PARAMETERS, RESULT> & {
     execute?: undefined;
   },
-): CoreTool<PARAMETERS, RESULT> & {
+): Tool<PARAMETERS, RESULT> & {
   execute: undefined;
 };
 export function tool<PARAMETERS extends Parameters, RESULT = any>(
-  tool: CoreTool<PARAMETERS, RESULT>,
-): CoreTool<PARAMETERS, RESULT> {
+  tool: Tool<PARAMETERS, RESULT>,
+): Tool<PARAMETERS, RESULT> {
   return tool;
 }