+++ packages/deepseek/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @ai-sdk/deepseek
 
+## 0.1.6
+
+### Patch Changes
+
+- Updated dependencies [3a602ca]
+  - @ai-sdk/provider-utils@2.1.5
+  - @ai-sdk/openai-compatible@0.1.6
+
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies [066206e]
+  - @ai-sdk/provider-utils@2.1.4
+  - @ai-sdk/openai-compatible@0.1.5
+
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies [39e5c1f]
+  - @ai-sdk/provider-utils@2.1.3
+  - @ai-sdk/openai-compatible@0.1.4
+
 ## 0.1.3
 
 ### Patch Changes